% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_SubTite.R
\name{Print_SubTite}
\alias{Print_SubTite}
\title{Gives summaries of GetSubTite Objects.}
\usage{
Print_SubTite(Z)
}
\arguments{
\item{Z}{List produced by GetSubTite.}
}
\description{
Gives summaries of GetSubTite Objects.
}
